SD flash接口形式封装

+++ src/driver/spi/drv_spi_sd_flash.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void    drv_spi_sd_flash_init(void);
-void    drv_spi_sd_flash_set_speed(uint8_t speed);
+void    drv_spi_sd_flash_set_speed(uint32_t speed);
 uint8_t drv_spi_sd_flash_transfer(uint8_t txdata);
 
 #endif //__DRV_SPI_SD_FLASH_H__
